+++ src/support.js
@@ -23,6 +23,17 @@
 			})
 		}
 
+		this.checkMongoId = function(mongoId){
+			var checkForHexRegExp = new RegExp("^[0-9a-fA-F]{24}$");
+			return syntaxValid = (((typeof mongoId) === 'string') && checkForHexRegExp.test(mongoId));
+		}
+
+		this.errorSender = function(status, message){
+			var deferred = Q.defer();
+			deferred.reject({"status":status, "message":message});
+			return deferred.promise
+		}
+
 	};
 	var support = new Support();
 	// Added Support Object to Stamplay